% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice.R
\name{Matrice}
\alias{Matrice}
\title{Title}
\usage{
Matrice(
  bp = 1000,
  chromosomelist,
  plasmidlist,
  k = 5,
  nchromosome = 1000,
  nplasmid = 1000
)
}
\arguments{
\item{nplasmid}{}
}
\value{

}
\description{
Title
}
